@@ -1,80 +1,79 @@
 ﻿using System;
 using Xunit;
 
-namespace Open.Text.Tests
+namespace Open.Text.Tests;
+
+public static class ComparableTests
 {
-	public static class ComparableTests
+	[Theory]
+	[InlineData("ABC", "abc")]
+	[InlineData("XyZ", "xYz")]
+	public static void CaseInsensitive(string a, string b)
 	{
-		[Theory]
-		[InlineData("ABC", "abc")]
-		[InlineData("XyZ", "xYz")]
-		public static void CaseInsensitive(string a, string b)
-		{
-			var cA = a.AsCaseInsensitive();
-			var cB = b.AsCaseInsensitive();
+		var cA = a.AsCaseInsensitive();
+		var cB = b.AsCaseInsensitive();
 
-			Assert.True(cA == cB);
-			Assert.True(cA == a);
-			Assert.True(cA == b);
-			Assert.True(cB == a);
-			Assert.True(cB == b);
+		Assert.True(cA == cB);
+		Assert.True(cA == a);
+		Assert.True(cA == b);
+		Assert.True(cB == a);
+		Assert.True(cB == b);
 
-			var sA = a.AsSpan().AsCaseInsensitive();
-			var sB = b.AsSpan().AsCaseInsensitive();
+		var sA = a.AsSpan().AsCaseInsensitive();
+		var sB = b.AsSpan().AsCaseInsensitive();
 
-			Assert.True(sA == sB);
-			Assert.True(sA == a);
-			Assert.True(sB == b);
-			Assert.True(sB == a);
-			Assert.True(sB == b);
+		Assert.True(sA == sB);
+		Assert.True(sA == a);
+		Assert.True(sB == b);
+		Assert.True(sB == a);
+		Assert.True(sB == b);
 
-			Assert.True(cA == sA);
-			Assert.True(cA == sB);
+		Assert.True(cA == sA);
+		Assert.True(cA == sB);
 
-			Assert.True(cB == sA);
-			Assert.True(cB == sB);
+		Assert.True(cB == sA);
+		Assert.True(cB == sB);
 
-			Assert.True(sA == cA);
-			Assert.True(sA == cB);
+		Assert.True(sA == cA);
+		Assert.True(sA == cB);
 
-			Assert.True(sB == cA);
-			Assert.True(sB == cB);
-		}
+		Assert.True(sB == cA);
+		Assert.True(sB == cB);
+	}
 
-		[Theory]
-		[InlineData("ABC", "abc")]
-		[InlineData("XyZ", "xYz")]
-		public static void CaseSensitive(string a, string b)
-		{
-			var cA = a.AsComparable(StringComparison.Ordinal);
-			var cB = b.AsComparable(StringComparison.Ordinal);
+	[Theory]
+	[InlineData("ABC", "abc")]
+	[InlineData("XyZ", "xYz")]
+	public static void CaseSensitive(string a, string b)
+	{
+		var cA = a.AsComparable(StringComparison.Ordinal);
+		var cB = b.AsComparable(StringComparison.Ordinal);
 
-			Assert.False(cA == cB);
-			Assert.True(cA == a);
-			Assert.False(cA == b);
-			Assert.False(cB == a);
-			Assert.True(cB == b);
+		Assert.False(cA == cB);	Assert.True(cA != cB);
+		Assert.True(cA == a);	Assert.False(cA != a);
+		Assert.False(cA == b);	Assert.True(cA != b);
+		Assert.False(cB == a);	Assert.True(cB != a);
+		Assert.True(cB == b);	Assert.False(cB != b);
 
-			var sA = a.AsSpan().AsComparable(StringComparison.Ordinal);
-			var sB = b.AsSpan().AsComparable(StringComparison.Ordinal);
+		var sA = a.AsSpan().AsComparable(StringComparison.Ordinal);
+		var sB = b.AsSpan().AsComparable(StringComparison.Ordinal);
 
-			Assert.False(sA == sB);
-			Assert.True(sA == a);
-			Assert.False(sA == b);
-			Assert.False(sB == a);
-			Assert.True(sB == b);
+		Assert.False(sA == sB); Assert.True(sA != sB);
+		Assert.True(sA == a);	Assert.False(sA != a);
+		Assert.False(sA == b);
+		Assert.False(sB == a);
+		Assert.True(sB == b);
 
-			Assert.True(cA == sA);
-			Assert.False(cA == sB);
+		Assert.True(cA == sA);
+		Assert.False(cA == sB);
 
-			Assert.False(cB == sA);
-			Assert.True(cB == sB);
+		Assert.False(cB == sA);
+		Assert.True(cB == sB);
 
-			Assert.True(sA == cA);
-			Assert.False(sA == cB);
+		Assert.True(sA == cA);
+		Assert.False(sA == cB);
 
-			Assert.False(sB == cA);
-			Assert.True(sB == cB);
-		}
+		Assert.False(sB == cA);
+		Assert.True(sB == cB);
 	}
 }
